# openapi/api-gateway.yaml
openapi: 3.0.3
info:
  title: DropMate API Gateway
  version: 0.1.0
  description: >
    Gateway-facing API contract aggregating DropMate identity and order services.
servers:
  - url: http://localhost:4000
    description: Local development via Docker Compose
paths:
  /healthz:
    get:
      summary: Gateway health probe
      responses:
        "200":
          description: Gateway is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  uptime:
                    type: number
  /api/v1/auth/register:
    post:
      summary: Register a new DropMate user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already registered
  /api/v1/auth/login:
    post:
      summary: Authenticate with email/password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /api/v1/auth/profile:
    get:
      summary: Retrieve the authenticated user's profile
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid or missing token
  /api/v1/orders:
    get:
      summary: List orders
      tags: [Orders]
      responses:
        "200":
          description: Orders fetched
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
    post:
      summary: Create an order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/orders/{orderId}:
    get:
      summary: Get an order by ID
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /api/v1/orders/{orderId}/assign:
    patch:
      summary: Assign a driver to an order
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDriverRequest"
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Order not found
  /api/v1/orders/{orderId}/status:
    patch:
      summary: Update order status
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Order not found
components:
  schemas:
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [manager, driver, customer]
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          format: password
        name:
          type: string
        role:
          type: string
          enum: [manager, driver, customer]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    Order:
      type: object
      required:
        [id, customerId, pickupAddress, dropoffAddress, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
        driverId:
          type: string
          nullable: true
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
        status:
          type: string
          enum: [created, assigned, picked_up, delivered, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [customerId, pickupAddress, dropoffAddress]
      properties:
        customerId:
          type: string
        pickupAddress:
          type: string
        dropoffAddress:
          type: string
    AssignDriverRequest:
      type: object
      required: [driverId]
      properties:
        driverId:
          type: string
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [created, assigned, picked_up, delivered, cancelled]
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT